task10/requirements.txt
fastapi~=0.108.0
uvicorn~=0.25.0
httpx
google-api-python-client
google-auth-oauthlib
python-dotenv

# linters
black
isort
